# BackEnd/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"roles": {
						"items": {},
						"type": "array"
					},
					"units": {
						"items": {},
						"type": "array"
					},
					"avatarIcon": {
						"type": "string"
					},
					"failedAttempts": {
						"type": "number",
						"format": "double"
					},
					"lockedUntil": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"username",
					"email",
					"roles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserCreationParams": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"units": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"username",
					"email",
					"roles"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateParams": {
				"properties": {
					"username": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"roles": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"units": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"avatarIcon": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"UnitResponse": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"translations": {
						"$ref": "#/components/schemas/Record_string.string_"
					}
				},
				"required": [
					"_id",
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnitCreationParams": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"translations": {
						"$ref": "#/components/schemas/Record_string.string_"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "IPT Form Builder API",
		"version": "1.0.0",
		"description": "Built by GP Team B 25/26:\n- Ângela Sebastião\n- Bernardo Ferreira\n- Catarina Agostinho\n- Joana Flores\n- José Fernandes\n- Paulo Pereira\n- Sady Neto",
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/UserResponse"
											},
											"type": "array"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden – admin role required"
					}
				},
				"description": "list existing users (admin only)",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "AddUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Missing parameters"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden – admin role required"
					},
					"409": {
						"description": "User already exists"
					}
				},
				"description": "create a new user (admin only).",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreationParams"
							}
						}
					}
				}
			}
		},
		"/users/{id}": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden – admin role required or modifying other user"
					},
					"404": {
						"description": "User not found"
					},
					"409": {
						"description": "Username or email already in use"
					}
				},
				"description": "Update user details (admin only)",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden – admin role required"
					},
					"404": {
						"description": "User not found"
					}
				},
				"description": "Delete a user by ID — soft delete (admin only)",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/units": {
			"get": {
				"operationId": "GetUnits",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UnitResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Units"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateUnit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UnitResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Name is required"
					},
					"409": {
						"description": "Unit already exists"
					}
				},
				"tags": [
					"Units"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UnitCreationParams"
							}
						}
					}
				}
			}
		},
		"/units/{id}": {
			"get": {
				"operationId": "GetUnit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UnitResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Unit not found"
					}
				},
				"tags": [
					"Units"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUnit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UnitResponse"
										},
										{
											"properties": {
												"message": {
													"type": "string"
												}
											},
											"required": [
												"message"
											],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "Unit not found"
					}
				},
				"tags": [
					"Units"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UnitCreationParams"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUnit",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Unit not found"
					}
				},
				"tags": [
					"Units"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}